Put messages in separate file
