+++ src/messages.js
@@ -1,0 +1,74 @@
+const messages = {
+  accounts: {
+    id: 'dapp.methods.accounts',
+    description: 'Explanation of the accounts methodGroup permission',
+    defaultMessage: 'Can view your Parity accounts'
+  },
+  accountsCreate: {
+    id: 'dapp.methods.accountsCreate',
+    description: 'Explanation of the accountsCreate methodGroup permission',
+    defaultMessage: 'Can create new Parity accounts'
+  },
+  accountsDelete: {
+    id: 'dapp.methods.accountsDelete',
+    description: 'Explanation of the accountsDelete methodGroup permission',
+    defaultMessage: 'Can delete your Parity accounts'
+  },
+  accountsEdit: {
+    id: 'dapp.methods.accountsEdit',
+    description: 'Explanation of the accountsEdit methodGroup permission',
+    defaultMessage: 'Can edit your Parity accounts'
+  },
+  dapps: {
+    id: 'dapp.methods.dapps',
+    description: 'Explanation of the dapps methodGroup permission',
+    defaultMessage: 'Can read info about dapps'
+  },
+  dappsEdit: {
+    id: 'dapp.methods.dappsEdit',
+    description: 'Explanation of the dappsEdit methodGroup permission',
+    defaultMessage: 'Can write info on dapps'
+  },
+  node: {
+    id: 'dapp.methods.node',
+    description: 'Explanation of the node methodGroup permission',
+    defaultMessage: 'Can read info about node'
+  },
+  nodeUpgrade: {
+    id: 'dapp.methods.nodeUpgrade',
+    description: 'Explanation of the nodeUpgrade methodGroup permission',
+    defaultMessage: 'Can upgrade Parity'
+  },
+  shell: {
+    id: 'dapp.methods.shell',
+    description: 'Explanation of the shell methodGroup permission',
+    defaultMessage: 'Can get access to the Shell'
+  },
+  signerConfirm: {
+    id: 'dapp.methods.signerConfirm',
+    description: 'Explanation of the signerConfirm methodGroup permission',
+    defaultMessage: 'Can confirm or reject requests from the signer'
+  },
+  signerRequests: {
+    id: 'dapp.methods.signerRequests',
+    description: 'Explanation of the signerRequests methodGroup permission',
+    defaultMessage: 'Can get access to signer requests'
+  },
+  vaults: {
+    id: 'dapp.methods.vaults',
+    description: 'Explanation of the vaults methodGroup permission',
+    defaultMessage: 'Can get access to your vaults'
+  },
+  vaultsCreate: {
+    id: 'dapp.methods.vaultsCreate',
+    description: 'Explanation of the vaultsCreate methodGroup permission',
+    defaultMessage: 'Can create new vaults'
+  },
+  vaultsEdit: {
+    id: 'dapp.methods.vaultsEdit',
+    description: 'Explanation of the vaultsEdit methodGroup permission',
+    defaultMessage: 'Can edit your vaults'
+  }
+};
+
+export default messages;
